Add Karpenter installation script for Kubernetes cluster
This addition includes a bash script for setting up Karpenter in a Kubernetes cluster. It contains the necessary details and steps to create an EKS cluster, set up Karpenter permissions, and includes the instructions to add the cluster to the kubeconfig. The script also provides a cleanup command to delete the cluster.